it's a final commit
In this i have made a tic tac toe game. in this i add index.html , index.css & script.js files. in this i make a program in which you can play a game and whenever you will win i have drop down a dancing cartoon gif which one will be displayed on the right side of the corner and if you click on the reset button it will be disappear. whenever you go in the any one of the box in the game then a colour of the box will be change into "aqua" color. morel of the game is to learn something new and make a game for play and  enjoyment....